review: split rewrites
